requirements.txt
airtable-python-wrapper==0.15.3
attrs==22.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
certifi==2022.9.24
cfscrape==2.1.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
exceptiongroup==1.1.0
greenlet==1.1.3
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
packaging==23.0
pexpect==4.8.0
Pillow==9.0.1
playwright==1.27.1
pluggy==1.0.0
plumbum==1.8.0
ptyprocess==0.7.0
pycairo==1.20.1
pyee==8.1.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-apt===2.4.0-elementary9-ubuntu7.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-slugify==8.0.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.12
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
